♻️Refactor: 메인컴포넌트, 모달컴포넌트 로직부분 커스텀훅 생성하여 분리
