+++ src/model/types/employeeInfo.type.ts
@@ -1,0 +1,10 @@
+export type EmployeeInfo = {
+  name: string;
+  email: string;
+  phoneNumber: string;
+  jobName: string;
+  salaryType: string;
+  salaryAmount: number;
+  companyCode: string;
+  uid?: string;
+};
